Add invoke-polymorphic-range instruction
Bug: 31543917

+++ dx/src/com/android/jack/dx/dex/code/form/Form4rcc.java
@@ -110,7 +110,7 @@ public final class Form4rcc extends InsnFormat {
     int count = regs.getWordCount();
     DualCstInsn dci = (DualCstInsn) insn;
 
-    write(out, opcodeUnit(insn, count), dci.getFirstIndex(), (short) firstReg,
+    write(out, opcodeUnit(insn, count), (short) dci.getFirstIndex(), (short) firstReg,
         (short) dci.getSecondIndex());
   }
 }
